requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
discord.py==2.4.0
distro==1.9.0
dnspython==2.7.0
faiss-cpu==1.7.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx>=0.25.2,<0.26.0
idna==3.10
jiter==0.8.2
multidict==6.1.0
numpy==2.2.3
ollama==0.1.6
packaging==24.2
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
psycopg2-binary==2.9.9
qdrant-client==1.7.0
openai==1.12.0
